Rename metrics-aggregator to router
