+++ components/router/config/default/kustomization.yaml
