+++ components/router/config/default/manager_auth_proxy_patch.yaml
