+++ components/router/config/default/manager_config_patch.yaml
