+++ components/router/config/manager/kustomization.yaml
