+++ components/router/config/rbac/auth_proxy_role.yaml
