+++ components/router/config/rbac/auth_proxy_service.yaml
