+++ components/router/config/rbac/kustomization.yaml
