+++ components/router/config/rbac/leader_election_role.yaml
