+++ components/router/hack/boilerplate.go.txt
