+++ components/router/pkg/backend/llamacpp.go
@@ -19,8 +19,8 @@ package backend
 import (
 	dto "github.com/prometheus/client_model/go"
 
-	"github.com/inftyai/metrics-aggregator/pkg/store"
-	"github.com/inftyai/metrics-aggregator/pkg/util"
+	"github.com/inftyai/router/pkg/store"
+	"github.com/inftyai/router/pkg/util"
 )
 
 var _ Backend = &LlamaCpp{}
