+++ components/router/pkg/backend/vllm.go
@@ -17,8 +17,8 @@ limitations under the License.
 package backend
 
 import (
-	"github.com/inftyai/metrics-aggregator/pkg/store"
-	"github.com/inftyai/metrics-aggregator/pkg/util"
+	"github.com/inftyai/router/pkg/store"
+	"github.com/inftyai/router/pkg/util"
 	dto "github.com/prometheus/client_model/go"
 )
