+++ components/router/pkg/dispatcher/plugins/latency-aware/score.go
@@ -19,8 +19,8 @@ package latencyAware
 import (
 	"context"
 
-	"github.com/inftyai/metrics-aggregator/pkg/dispatcher/framework"
-	"github.com/inftyai/metrics-aggregator/pkg/store"
+	"github.com/inftyai/router/pkg/dispatcher/framework"
+	"github.com/inftyai/router/pkg/store"
 )
 
 var _ framework.ScorePlugin = &LatencyAware{}
